+++ src/main/kotlin/com/github/rafaelldi/tyeplugin/settings/TyeSettingsConfigurable.kt
@@ -13,9 +13,9 @@ class TyeSettingsConfigurable(private val project: Project) : Configurable {
 
     override fun isModified(): Boolean {
         val settings = TyeSettingsState.getInstance(project)
-        var modified = component.getTyeToolPath() != settings.tyeToolPath
-        modified = modified or (component.getOverwriteTyeFile() != settings.overwriteTyeFile)
-        return modified
+        return (component.getTyeToolPath() != settings.tyeToolPath) or
+                (component.getOverwriteTyeFile() != settings.overwriteTyeFile) or
+                (component.getCheckTyeNewVersions() != settings.checkTyeNewVersions)
     }
 
     override fun apply() {
@@ -23,6 +23,7 @@ class TyeSettingsConfigurable(private val project: Project) : Configurable {
         settings.tyeToolPath = component.getTyeToolPath()
         settings.tyeHost = component.getTyeHost()
         settings.overwriteTyeFile = component.getOverwriteTyeFile()
+        settings.checkTyeNewVersions = component.getCheckTyeNewVersions()
     }
 
     override fun reset() {
@@ -30,6 +31,7 @@ class TyeSettingsConfigurable(private val project: Project) : Configurable {
         component.setTyeToolPath(settings.tyeToolPath)
         component.setTyeHost(settings.tyeHost)
         component.setOverwriteTyeFile(settings.overwriteTyeFile)
+        component.setCheckTyeNewVersions(settings.checkTyeNewVersions)
     }
 
     override fun getDisplayName(): String = "Tye"