+++ src/main/kotlin/com/github/rafaelldi/tyeplugin/settings/TyeSettingsState.kt
@@ -1,10 +1,7 @@
 package com.github.rafaelldi.tyeplugin.settings
 
 import com.github.rafaelldi.tyeplugin.services.TyeGlobalToolPathProvider
-import com.intellij.openapi.components.PersistentStateComponent
-import com.intellij.openapi.components.ServiceManager
-import com.intellij.openapi.components.State
-import com.intellij.openapi.components.Storage
+import com.intellij.openapi.components.*
 import com.intellij.openapi.project.Project
 import com.intellij.util.xmlb.XmlSerializerUtil
 
@@ -15,13 +12,13 @@ import com.intellij.util.xmlb.XmlSerializerUtil
 class TyeSettingsState : PersistentStateComponent<TyeSettingsState> {
     companion object {
         @JvmStatic
-        fun getInstance(project: Project): TyeSettingsState =
-            ServiceManager.getService(project, TyeSettingsState::class.java)
+        fun getInstance(project: Project): TyeSettingsState = project.service()
     }
 
     var tyeToolPath = TyeGlobalToolPathProvider.findDefaultTyeGlobalToolPath()
     var tyeHost = "http://localhost:8000"
     var overwriteTyeFile = false
+    var checkTyeNewVersions = false
 
     override fun getState(): TyeSettingsState = this