+++ src/main/kotlin/com/github/rafaelldi/tyeplugin/startup/TyeUpdateGlobalToolStartupActivity.kt
@@ -1,7 +1,9 @@
 package com.github.rafaelldi.tyeplugin.startup
 
+import com.github.rafaelldi.tyeplugin.actions.DisableTyeNewVersionCheckAction
 import com.github.rafaelldi.tyeplugin.actions.UpdateTyeGlobalToolNotificationAction
 import com.github.rafaelldi.tyeplugin.services.TyeGlobalToolService
+import com.github.rafaelldi.tyeplugin.settings.TyeSettingsState
 import com.intellij.notification.Notification
 import com.intellij.notification.NotificationType
 import com.intellij.openapi.components.service
@@ -11,6 +13,10 @@ import com.intellij.openapi.startup.StartupActivity
 
 class TyeUpdateGlobalToolStartupActivity : StartupActivity, DumbAware {
     override fun runActivity(project: Project) {
+        val settings = TyeSettingsState.getInstance(project)
+        if (!settings.checkTyeNewVersions)
+            return
+
         val tyeGlobalToolService = project.service<TyeGlobalToolService>()
         val isActualVersionInstalled = tyeGlobalToolService.isActualTyeGlobalToolVersionInstalled()
 
@@ -18,6 +24,7 @@ class TyeUpdateGlobalToolStartupActivity : StartupActivity, DumbAware {
 
         Notification("Tye", "New version of tye global tool is available", "", NotificationType.INFORMATION)
             .addAction(UpdateTyeGlobalToolNotificationAction())
+            .addAction(DisableTyeNewVersionCheckAction())
             .notify(project)
     }
 }