+++ src/main/kotlin/com/github/rafaelldi/tyeplugin/util/File.kt
@@ -4,4 +4,4 @@ import com.intellij.openapi.vfs.VirtualFile
 
 const val TYE_FILE_NAME = "tye.yaml"
 
-fun VirtualFile?.isTyeFile() = this?.name?.toLowerCase() == TYE_FILE_NAME
+fun VirtualFile?.isTyeFile() = this?.name?.lowercase() == TYE_FILE_NAME